web/requirements.txt
fastapi
uvicorn
paho-mqtt
sqlalchemy
psycopg2-binary